+++ src/faebryk/exporters/pcb/kicad/transformer.py
@@ -3,14 +3,12 @@
 
 import itertools
 import logging
-import math
 import pprint
 import random
 import re
 from operator import add
 from typing import Any, TypeVar
 
-import numpy as np
 from faebryk.core.core import (
     Footprint as Core_Footprint,
 )
@@ -26,6 +24,7 @@ from faebryk.library.has_kicad_footprint import has_kicad_footprint
 from faebryk.library.has_overriden_name import has_overriden_name
 from faebryk.library.has_pcb_position import has_pcb_position
 from faebryk.library.Net import Net as FNet
+from faebryk.libs.geometry.basic import Geometry
 from faebryk.libs.kicad.pcb import (
     PCB,
     Arc,
@@ -122,141 +121,6 @@ class PCB_Transformer:
         }
         logger.debug(f"Attached: {pprint.pformat(attached)}")
 
-    def set_dimensions_rectangle(
-        self,
-        width_mm: float,
-        height_mm: float,
-        origin_x_mm: float = 0.0,
-        origin_y_mm: float = 0.0,
-    ):
-        """
-        Create a rectengular board outline (edge cut)
-        """
-
-        for geo in self.pcb.geoms:
-            if not isinstance(geo, GR_Line):
-                continue
-            line = geo
-            if line.layer_name != "Edge.Cuts":
-                continue
-            line.delete()
-
-        points = [
-            (origin_x_mm, origin_y_mm),
-            (origin_x_mm, origin_y_mm + height_mm),
-            (origin_x_mm + width_mm, origin_y_mm + height_mm),
-            (origin_x_mm + width_mm, origin_y_mm),
-            (origin_x_mm, origin_y_mm),
-        ]
-
-        for start, end in zip(points[:-1], points[1:]):
-            self.pcb.append(
-                GR_Line.factory(
-                    start,
-                    end,
-                    stroke=GR_Line.Stroke.factory(0.05, "default"),
-                    layer="Edge.Cuts",
-                    tstamp=str(int(random.random() * 100000)),
-                )
-            )
-
-        self.dimensions = (width_mm, height_mm)
-
-    def set_dimensions_complex(
-        self,
-        points: list,
-        origin_x_mm: float = 0.0,
-        origin_y_mm: float = 0.0,
-        remove_existing_outline: bool = True,
-    ):
-        """
-        Create a board outline (edge cut) consisting out of
-        points connected via straight lines
-        """
-
-        # remove existing lines on Egde.cuts layer
-        if remove_existing_outline:
-            for geo in self.pcb.geoms:
-                if not isinstance(geo, GR_Line):
-                    continue
-                line = geo
-                if line.layer_name != "Edge.Cuts":
-                    continue
-                line.delete()
-
-        # offset the points with origin
-        points_offset = []
-        for point in points:
-            points_offset.append((point[0] + origin_x_mm, point[1] + origin_y_mm))
-
-        # Append the first point to the end to create a closed shape
-        points_offset.append(points_offset[0])
-
-        # create Edge.Cuts lines
-        for start, end in zip(points_offset, points_offset[1:]):
-            self.pcb.append(
-                GR_Line.factory(
-                    start,
-                    end,
-                    stroke=GR_Line.Stroke.factory(0.05, "default"),
-                    layer="Edge.Cuts",
-                    tstamp=str(int(random.random() * 100000)),
-                )
-            )
-
-        width_mm = max(p[0] for p in points)
-        height_mm = max(p[1] for p in points)
-
-        self.dimensions = (width_mm, height_mm)
-
-    def move_fp(self, fp: Footprint, coord: Footprint.Coord, layer: str):
-        if any([x.text == "FBRK:notouch" for x in fp.user_text]):
-            logger.warning(f"Skipped no touch component: {fp.name}")
-            return
-
-        # Rotate
-        rot_angle = (coord[2] - fp.at.coord[2]) % 360
-
-        if rot_angle:
-            for at_prop in fp.get_prop("at", recursive=True):
-                coords = at_prop.node[1:]
-                # if rot is 0 in coord, its compressed to a 2-tuple by kicad
-                if len(coords) == 2:
-                    coords.append(0)
-                coords[2] = (coords[2] + rot_angle) % 360
-                at_prop.node[1:] = coords
-
-        fp.at.coord = coord
-
-        # Flip
-        # TODO a bit ugly with the hardcoded front layer
-        flip = layer != "F.Cu" and fp.layer != layer
-
-        if flip:
-            for layer_prop in fp.get_prop(["layer", "layers"], recursive=True):
-
-                def _flip(x):
-                    return (
-                        x.replace("F.", "<F>.")
-                        .replace("B.", "F.")
-                        .replace("<F>.", "B.")
-                    )
-
-                layer_prop.node[1:] = [_flip(x) for x in layer_prop.node[1:]]
-
-        # Label
-        if any([x.text == "FBRK:autoplaced" for x in fp.user_text]):
-            return
-        fp.append(
-            FP_Text.factory(
-                text="FBRK:autoplaced",
-                at=At.factory((0, 0, 0)),
-                font=self.font,
-                tstamp=str(next(self.tstamp_i)),
-                layer="User.5",
-            )
-        )
-
     def cleanup(self):
         # delete auto-placed vias
         # determined by their size_drill values
@@ -276,6 +140,16 @@ class PCB_Transformer:
             geo.delete()
         self.pcb.garbage_collect()
 
+    T = TypeVar("T")
+
+    @staticmethod
+    def flipped(input_list: list[tuple[T, int]]) -> list[tuple[T, int]]:
+        return [(x, (y + 180) % 360) for x, y in reversed(input_list)]
+
+    def gen_tstamp(self):
+        return str(next(self.tstamp_i))
+
+    # Getter ---------------------------------------------------------------------------
     @staticmethod
     def get_fp(cmp) -> Footprint:
         return cmp.get_trait(PCB_Transformer.has_linked_kicad_footprint).get_fp()
@@ -295,7 +169,7 @@ class PCB_Transformer:
                 lines = [(geo.start, geo.end)]
             elif isinstance(geo, Arc):
                 arc = (geo.start, geo.mid, geo.end)
-                lines = self.Geometry.approximate_arc(*arc, resolution=10)
+                lines = Geometry.approximate_arc(*arc, resolution=10)
             elif isinstance(geo, Rect):
                 rect = (geo.start, geo.end)
 
@@ -314,7 +188,7 @@ class PCB_Transformer:
             if geo.sym.startswith("fp"):
                 assert isinstance(parent, Footprint)
                 lines = [
-                    tuple(self.Geometry.abs_pos(parent.at.coord, x) for x in line)
+                    tuple(Geometry.abs_pos(parent.at.coord, x) for x in line)
                     for line in lines
                 ]
 
@@ -372,15 +246,68 @@ class PCB_Transformer:
 
         return get_geometry(polys, 0).exterior.coords
 
-    T = TypeVar("T")
+    @staticmethod
+    def get_corresponding_fp(
+        intf: ModuleInterface,
+    ) -> tuple[Core_Footprint, Module]:
+        obj = intf
+
+        while not obj.has_trait(has_footprint):
+            parent = obj.get_parent()
+            if parent is None:
+                raise Exception
+            obj = parent[0]
+
+        assert isinstance(obj, Module)
+        return obj.get_trait(has_footprint).get_footprint(), obj
 
     @staticmethod
-    def flipped(input_list: list[tuple[T, int]]) -> list[tuple[T, int]]:
-        return [(x, (y + 180) % 360) for x, y in reversed(input_list)]
+    def get_pad(intf: ModuleInterface) -> tuple[Footprint, Pad]:
+        cfp, obj = PCB_Transformer.get_corresponding_fp(intf)
+        pin_map = cfp.get_trait(has_kicad_footprint).get_pin_names()
+        cfg_if = [
+            (pin, name) for pin, name in pin_map.items() if intf.is_connected_to(pin)
+        ]
+        assert len(cfg_if) == 1
 
-    def gen_tstamp(self):
-        return str(next(self.tstamp_i))
+        pin_name = cfg_if[0][1]
 
+        fp = PCB_Transformer.get_fp(obj)
+        pad = fp.get_pad(pin_name)
+
+        return fp, pad
+
+    def get_copper_layers(self):
+        COPPER = re.compile(r"^.*\.Cu$")
+
+        return {name for name in self.pcb.layer_names if COPPER.match(name) is not None}
+
+    def get_copper_layers_pad(self, pad: Pad):
+        COPPER = re.compile(r"^.*\.Cu$")
+
+        all_layers = self.pcb.layer_names
+
+        def dewildcard(layer: str):
+            if "*" not in layer:
+                return {layer}
+            pattern = re.compile(layer.replace(".", r"\.").replace("*", r".*"))
+            return {
+                global_layer
+                for global_layer in all_layers
+                if pattern.match(global_layer) is not None
+            }
+
+        layers = pad.layers
+        dewildcarded_layers = {
+            sub_layer for layer in layers for sub_layer in dewildcard(layer)
+        }
+        matching_layers = {
+            layer for layer in dewildcarded_layers if COPPER.match(layer) is not None
+        }
+
+        return matching_layers
+
+    # Insert ---------------------------------------------------------------------------
     def insert(self, node: PCB_Node):
         self.pcb.append(node)
 
@@ -457,72 +384,11 @@ class PCB_Transformer:
     def insert_geo(self, geo: Geom):
         self.insert(geo)
 
-    @staticmethod
-    def get_corresponding_fp(
-        intf: ModuleInterface,
-    ) -> tuple[Core_Footprint, Module]:
-        obj = intf
-
-        while not obj.has_trait(has_footprint):
-            parent = obj.get_parent()
-            if parent is None:
-                raise Exception
-            obj = parent[0]
-
-        assert isinstance(obj, Module)
-        return obj.get_trait(has_footprint).get_footprint(), obj
-
-    @staticmethod
-    def get_pad(intf: ModuleInterface) -> tuple[Footprint, Pad]:
-        cfp, obj = PCB_Transformer.get_corresponding_fp(intf)
-        pin_map = cfp.get_trait(has_kicad_footprint).get_pin_names()
-        cfg_if = [
-            (pin, name) for pin, name in pin_map.items() if intf.is_connected_to(pin)
-        ]
-        assert len(cfg_if) == 1
-
-        pin_name = cfg_if[0][1]
-
-        fp = PCB_Transformer.get_fp(obj)
-        pad = fp.get_pad(pin_name)
-
-        return fp, pad
-
-    def get_copper_layers(self):
-        COPPER = re.compile(r"^.*\.Cu$")
-
-        return {name for name in self.pcb.layer_names if COPPER.match(name) is not None}
-
-    def get_copper_layers_pad(self, pad: Pad):
-        COPPER = re.compile(r"^.*\.Cu$")
-
-        all_layers = self.pcb.layer_names
-
-        def dewildcard(layer: str):
-            if "*" not in layer:
-                return {layer}
-            pattern = re.compile(layer.replace(".", r"\.").replace("*", r".*"))
-            return {
-                global_layer
-                for global_layer in all_layers
-                if pattern.match(global_layer) is not None
-            }
-
-        layers = pad.layers
-        dewildcarded_layers = {
-            sub_layer for layer in layers for sub_layer in dewildcard(layer)
-        }
-        matching_layers = {
-            layer for layer in dewildcarded_layers if COPPER.match(layer) is not None
-        }
-
-        return matching_layers
-
     def insert_via_next_to(self, intf: ModuleInterface, clearance: tuple[float, float]):
         fp, pad = self.get_pad(intf)
 
         rel_target = tuple(map(add, pad.at.coord, clearance))
-        coord = self.Geometry.abs_pos(fp.at.coord, rel_target)
+        coord = Geometry.abs_pos(fp.at.coord, rel_target)
 
         self.insert_via(coord[:2], pad.net)
 
@@ -545,15 +411,15 @@ class PCB_Transformer:
         start = pads[0].at.coord
 
         # construct triangle
-        shape = self.Geometry.triangle(
-            self.Geometry.abs_pos(fp.at.coord, start),
+        shape = Geometry.triangle(
+            Geometry.abs_pos(fp.at.coord, start),
             width=width,
             depth=depth,
             count=len(pads),
         )
 
         # clearance
-        shape = self.Geometry.translate(
+        shape = Geometry.translate(
             tuple([clearance if x != 0 else 0 for x in rect]), shape
         )
 
@@ -576,22 +442,22 @@ class PCB_Transformer:
 
         # from first & last pad
         start = pads[0].at.coord
-        abs_start = self.Geometry.abs_pos(fp.at.coord, start)
+        abs_start = Geometry.abs_pos(fp.at.coord, start)
 
-        shape = self.Geometry.line(
+        shape = Geometry.line(
             start=abs_start,
             length=length,
             count=len(pads),
         )
 
-        shape = self.Geometry.rotate(
+        shape = Geometry.rotate(
             axis=abs_start[:2],
             structure=shape,
             angle_deg=angle_deg,
         )
 
         # clearance
-        shape = self.Geometry.translate((clearance, 0), shape)
+        shape = Geometry.translate((clearance, 0), shape)
 
         # place vias
         for pad, point in zip(pads, shape):
@@ -610,11 +476,11 @@ class PCB_Transformer:
 
         # from first & last pad
         start = tuple(map(add, pads[0].at.coord, clearance))
-        abs_start = self.Geometry.abs_pos(fp.at.coord, start)
+        abs_start = Geometry.abs_pos(fp.at.coord, start)
 
-        shape = self.Geometry.line2(
+        shape = Geometry.line2(
             start=abs_start,
-            end=self.Geometry.abs_pos(abs_start, length),
+            end=Geometry.abs_pos(abs_start, length),
             count=len(pads),
         )
 
@@ -622,7 +488,7 @@ class PCB_Transformer:
         for pad, point in zip(pads, shape):
             self.insert_via(point, pad.net)
 
-    # Positioning ------------------------------------------------------------
+    # Positioning ----------------------------------------------------------------------
     def move_footprints(self):
         # position modules with defined positions
         pos_mods: set[Module] = {
@@ -647,183 +513,125 @@ class PCB_Transformer:
             logger.debug(f"Placing {fp.name} at {coord} layer {layer_name[coord[3]]}")
             self.move_fp(fp, coord[:3], layer_name[coord[3]])
 
-    # Geometry ----------------------------------------------------------------
-    class Geometry:
-        Point2D = tuple[float, float]
-        # TODO fix all Point2D functions to use Point
-        Point = has_pcb_position.Point
+    def move_fp(self, fp: Footprint, coord: Footprint.Coord, layer: str):
+        if any([x.text == "FBRK:notouch" for x in fp.user_text]):
+            logger.warning(f"Skipped no touch component: {fp.name}")
+            return
 
-        @staticmethod
-        def mirror(axis: tuple[float | None, float | None], structure: list[Point2D]):
-            return [
-                (
-                    2 * axis[0] - x if axis[0] is not None else x,
-                    2 * axis[1] - y if axis[1] is not None else y,
+        # Rotate
+        rot_angle = (coord[2] - fp.at.coord[2]) % 360
+
+        if rot_angle:
+            for at_prop in fp.get_prop("at", recursive=True):
+                coords = at_prop.node[1:]
+                # if rot is 0 in coord, its compressed to a 2-tuple by kicad
+                if len(coords) == 2:
+                    coords.append(0)
+                coords[2] = (coords[2] + rot_angle) % 360
+                at_prop.node[1:] = coords
+
+        fp.at.coord = coord
+
+        # Flip
+        # TODO a bit ugly with the hardcoded front layer
+        flip = layer != "F.Cu" and fp.layer != layer
+
+        if flip:
+            for layer_prop in fp.get_prop(["layer", "layers"], recursive=True):
+
+                def _flip(x):
+                    return (
+                        x.replace("F.", "<F>.")
+                        .replace("B.", "F.")
+                        .replace("<F>.", "B.")
+                    )
+
+                layer_prop.node[1:] = [_flip(x) for x in layer_prop.node[1:]]
+
+        # Label
+        if any([x.text == "FBRK:autoplaced" for x in fp.user_text]):
+            return
+        fp.append(
+            FP_Text.factory(
+                text="FBRK:autoplaced",
+                at=At.factory((0, 0, 0)),
+                font=self.font,
+                tstamp=str(next(self.tstamp_i)),
+                layer="User.5",
+            )
+        )
+
+    # Edge -----------------------------------------------------------------------------
+    def set_dimensions_rectangle(
+        self,
+        width_mm: float,
+        height_mm: float,
+        origin_x_mm: float = 0.0,
+        origin_y_mm: float = 0.0,
+        remove_existing_outline: bool = True,
+    ):
+        """
+        Create a rectengular board outline (edge cut)
+        """
+
+        points = [
+            (origin_x_mm, origin_y_mm),
+            (origin_x_mm, origin_y_mm + height_mm),
+            (origin_x_mm + width_mm, origin_y_mm + height_mm),
+            (origin_x_mm + width_mm, origin_y_mm),
+            (origin_x_mm, origin_y_mm),
+        ]
+
+        self.set_dimensions_complex(
+            points=points,
+            origin_x_mm=origin_x_mm,
+            origin_y_mm=origin_y_mm,
+            remove_existing_outline=remove_existing_outline,
+        )
+
+    def set_dimensions_complex(
+        self,
+        points: list,
+        origin_x_mm: float = 0.0,
+        origin_y_mm: float = 0.0,
+        remove_existing_outline: bool = True,
+    ):
+        """
+        Create a board outline (edge cut) consisting out of
+        points connected via straight lines
+        """
+
+        # remove existing lines on Egde.cuts layer
+        if remove_existing_outline:
+            for geo in self.pcb.geoms:
+                if not isinstance(geo, GR_Line):
+                    continue
+                line = geo
+                if line.layer_name != "Edge.Cuts":
+                    continue
+                line.delete()
+
+        # offset the points with origin
+        points_offset = []
+        for point in points:
+            points_offset.append((point[0] + origin_x_mm, point[1] + origin_y_mm))
+
+        # Append the first point to the end to create a closed shape
+        points_offset.append(points_offset[0])
+
+        # create Edge.Cuts lines
+        for start, end in zip(points_offset, points_offset[1:]):
+            self.pcb.append(
+                GR_Line.factory(
+                    start,
+                    end,
+                    stroke=GR_Line.Stroke.factory(0.05, "default"),
+                    layer="Edge.Cuts",
+                    tstamp=str(int(random.random() * 100000)),
                 )
-                for (x, y) in structure
-            ]
-
-        @staticmethod
-        def abs_pos(parent: Point, child: Point) -> Point:
-            rot_deg = parent[2] + child[2]
-
-            rot = rot_deg / 360 * 2 * math.pi
-
-            x, y = parent[:2]
-            cx, cy = child[:2]
-
-            rx = round(cx * math.cos(rot) + cy * math.sin(rot), 2)
-            ry = round(-cx * math.sin(rot) + cy * math.cos(rot), 2)
-
-            # print(f"Rotate {round(cx,2),round(cy,2)},
-            # by {round(rot,2),parent[2]} to {rx,ry}")
-
-            # TODO not sure what this is supposed to do
-            # for i in range(2, len(parent)):
-            #    if len(child) <= i:
-            #        continue
-            #    if parent[i] != 0 and child[i] != 0:
-            #        logger.warn(f"Adding non-zero values: {parent[i]=} + {child[i]=}")
-
-            # TODO check if this works everywhere
-            out = (
-                # XY
-                x + rx,
-                y + ry,
-                # ROT
-                *(c1 + c2 for c1, c2 in zip(parent[2:3], child[2:3])),
-                # Layer
-                *child[3:],
             )
 
-            return out
+        width_mm = max(p[0] for p in points)
+        height_mm = max(p[1] for p in points)
 
-        @staticmethod
-        def translate(vec: Point2D, structure: list[Point2D]) -> list[Point2D]:
-            return [tuple(map(add, vec, point)) for point in structure]
-
-        @classmethod
-        def rotate(
-            cls, axis: Point2D, structure: list[Point2D], angle_deg: float
-        ) -> list[Point2D]:
-            theta = np.radians(angle_deg)
-            c, s = np.cos(theta), np.sin(theta)
-            R = np.array(((c, -s), (s, c)))
-
-            return cls.translate(
-                (-axis[0], -axis[1]),
-                [
-                    tuple(R @ np.array(point))
-                    for point in cls.translate(axis, structure)
-                ],
-            )
-
-        C = TypeVar("C", tuple[float, float], tuple[float, float, float])
-
-        @staticmethod
-        def triangle(start: C, width: float, depth: float, count: int):
-            x1, y1 = start[:2]
-
-            n = count - 1
-            cy = width / n
-
-            ys = [round(y1 + cy * i, 2) for i in range(count)]
-            xs = [
-                round(x1 + depth * (1 - abs(1 - 1 / n * i * 2)), 2)
-                for i in range(count)
-            ]
-
-            return list(zip(xs, ys))
-
-        @staticmethod
-        def line(start: C, length: float, count: int):
-            x1, y1 = start[:2]
-
-            n = count - 1
-            cy = length / n
-
-            ys = [round(y1 + cy * i, 2) for i in range(count)]
-            xs = [x1] * count
-
-            return list(zip(xs, ys))
-
-        @staticmethod
-        def line2(start: C, end: C, count: int):
-            x1, y1 = start[:2]
-            x2, y2 = end[:2]
-
-            n = count - 1
-            cx = (x2 - x1) / n
-            cy = (y2 - y1) / n
-
-            ys = [round(y1 + cy * i, 2) for i in range(count)]
-            xs = [round(x1 + cx * i, 2) for i in range(count)]
-
-            return list(zip(xs, ys))
-
-        @staticmethod
-        def find_circle_center(p1, p2, p3):
-            """
-            Finds the center of the circle passing through the three given points.
-            """
-            # Compute the midpoints
-            mid1 = (p1 + p2) / 2
-            mid2 = (p2 + p3) / 2
-
-            # Compute the slopes of the lines
-            m1 = (p2[1] - p1[1]) / (p2[0] - p1[0])
-            m2 = (p3[1] - p2[1]) / (p3[0] - p2[0])
-
-            # The slopes of the perpendicular bisectors
-            perp_m1 = -1 / m1
-            perp_m2 = -1 / m2
-
-            # Equations of the perpendicular bisectors
-            # y = perp_m1 * (x - mid1[0]) + mid1[1]
-            # y = perp_m2 * (x - mid2[0]) + mid2[1]
-
-            # Solving for x
-            x = (mid2[1] - mid1[1] + perp_m1 * mid1[0] - perp_m2 * mid2[0]) / (
-                perp_m1 - perp_m2
-            )
-
-            # Solving for y using one of the bisector equations
-            y = perp_m1 * (x - mid1[0]) + mid1[1]
-
-            return np.array([x, y])
-
-        @staticmethod
-        def approximate_arc(p_start, p_mid, p_end, resolution=10):
-            p_start, p_mid, p_end = (np.array(p) for p in (p_start, p_mid, p_end))
-
-            # Calculate the center of the circle
-            center = PCB_Transformer.Geometry.find_circle_center(p_start, p_mid, p_end)
-
-            # Calculate start, mid, and end angles
-            start_angle = np.arctan2(p_start[1] - center[1], p_start[0] - center[0])
-            mid_angle = np.arctan2(p_mid[1] - center[1], p_mid[0] - center[0])
-            end_angle = np.arctan2(p_end[1] - center[1], p_end[0] - center[0])
-
-            # Adjust angles if necessary
-            if start_angle > mid_angle:
-                start_angle -= 2 * np.pi
-            if mid_angle > end_angle:
-                mid_angle -= 2 * np.pi
-
-            # Radius of the circle
-            r = np.linalg.norm(p_start - center)
-
-            # Compute angles of line segment endpoints
-            angles = np.linspace(start_angle, end_angle, resolution + 1)
-
-            # Compute points on the arc
-            points = np.array(
-                [[center[0] + r * np.cos(a), center[1] + r * np.sin(a)] for a in angles]
-            )
-
-            # Create line segments
-            segments = [(points[i], points[i + 1]) for i in range(resolution)]
-
-            seg_no_np = [(tuple(a), tuple(b)) for a, b in segments]
-
-            return seg_no_np
+        self.dimensions = (width_mm, height_mm)
